refactored signup api into routes/users.js route handler.
