--- backend-api/index.js
@@ -1,38 +1,0 @@
-// const createError = require('http-errors');
-const express = require('express');
-// const path = require('path');
-// const cookieParser = require('cookie-parser');
-// const logger = require('morgan');
-// const { sequelize } = require('./db/models');
-// const session = require('express-session');
-// const SequelizeStore = require('connect-session-sequelize')(session.Store);
-const usersRouter = require('./routes/users');
-const router = express.Router();
-
-// // view engine setup
-// app.set('view engine', 'pug');
-
-// app.use(logger('dev'));
-// app.use(express.json());
-// app.use(express.urlencoded({ extended: false }));
-// app.use(cookieParser());
-// app.use(express.static(path.join(__dirname, 'public')));
-
-// // set up session middleware
-// const store = new SequelizeStore({ db: sequelize });
-
-// app.use(
-//     session({
-//         secret: 'superSecret',
-//         store,
-//         saveUninitialized: false,
-//         resave: false,
-//     })
-// );
-
-// create Session table if it doesn't already exist
-// store.sync();
-
-router.use('/users', usersRouter);
-
-module.exports = router;
