+++ routes/users.js
@@ -1,21 +1,57 @@
-var express = require('express');
-var router = express.Router();
-const { csrfProtection, asyncHandler, } = require('../utils');
+const express = require('express');
+const bcrypt = require('bcryptjs');
+const { validationResult } = require('express-validator');
+const db = require('../db/models');
+const { csrfProtection, asyncHandler, userValidators } = require('../utils');
+const { loginUser } = require('../auth');
 
-/* GET users listing. */
-router.get('/', function (req, res, next) {
-  res.send('respond with a resource');
-});
+const router = express.Router();
 
-//route for registration (front-end)
+// get /signup
 router.get('/signup', csrfProtection, asyncHandler(async (req, res) => {
   res.render('user-signup', {
     title: 'Sign Up',
     csrfToken: req.csrfToken(),
-  })
-}))
+  });
 
-//route for registration (back-end)
-// router.post('/signup', )
+  console.log(res.locals.error)
+}));
+
+// post /signup
+router.post('/signup', csrfProtection, userValidators, asyncHandler(async (req, res) => {
+  const { username, firstName, lastName, email, password } = req.body;
+  const user = await db.User.build({
+    username,
+    firstName,
+    lastName,
+    email
+  });
+
+  const validatorErrors = validationResult(req);
+
+  if (validatorErrors.isEmpty()) {
+    const hashPass = await bcrypt.hash(password, 10);
+    user.hashedPassword = hashPass;
+    await user.save();
+
+    loginUser(req, res, user);
+
+    res.redirect('/');
+  } else {
+    const errors = validatorErrors.array().map(error => error.msg);
+
+    res.render('user-signup', {
+      title: "Sign Up",
+      email,
+      username,
+      firstName,
+      lastName,
+      csrfToken: req.csrfToken(),
+      errors
+    });
+
+  }
+
+}));
 
 module.exports = router;
