+++ views/user-signup.pug
@@ -1,0 +1,15 @@
+extends layout.pug
+
+include utils.pug
+
+block content
+    +errorSummary(errors)
+    form(action='/users/signup' method='post')
+        input(type='hidden' name='_csrf' value=csrfToken)
+        +field('Username', 'username', username)
+        +field('First Name', 'firstName', firstName)
+        +field('Last Name', 'lastName', lastName)
+        +field('Email Address', 'email', email, 'email')
+        +field('Password', 'password', password, 'password')
+        +field('Confirm Password', 'confirmPassword', confirmPassword, 'password')
+        button(type='submit') Sign Up
